refactor: move rbac_permission_required decorator to controllers/common/wraps
- Move rbac_permission_required and its helper functions (_is_resource_owned_by_current_user, _extract_resource_id) from controllers/console/wraps.py to controllers/common/wraps.py
- Update all import statements across the codebase to import rbac_permission_required from controllers.common.wraps instead of controllers.console.wraps
- Remove unused Literal import from console/wraps.py
- This makes the decorator available to all controllers, not just console-specific ones
